refactor(global): 외부 consumer import 수정 및 기존 파일 삭제
- `OpenTicketApplication`, `BookingController`, `SeatController` 등 9개 파일 import 경로 변경
- `global/queue/` 디렉토리 삭제
- `global/config/RedissonConfig.java` 삭제
- `resources/lua/queue_*.lua` 원본 삭제
