+++ src/main/java/com/example/openticket/api/service/queue/dto/response/QueueStatusResponse.java
@@ -1,7 +1,7 @@
 package com.example.openticket.api.service.queue.dto.response;
 
-import com.example.openticket.global.queue.QueuePhase;
-import com.example.openticket.global.queue.QueueStatus;
+import com.example.openticket.domain.queue.QueuePhase;
+import com.example.openticket.domain.queue.QueueStatus;
 
 public record QueueStatusResponse(
         String token,
