+++ src/main/java/com/example/openticket/global/config/WebConfig.java
@@ -2,7 +2,7 @@ package com.example.openticket.global.config;
 
 import com.example.openticket.global.auth.AuthenticationInterceptor;
 import com.example.openticket.global.auth.LoginUserArgumentResolver;
-import com.example.openticket.global.queue.QueueValidationInterceptor;
+import com.example.openticket.global.interceptor.QueueValidationInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
