--- src/main/java/com/example/openticket/global/queue/ActiveToken.java
@@ -1,7 +1,0 @@
-package com.example.openticket.global.queue;
-
-public record ActiveToken(
-        Long userId,
-        long expiresAt
-) {
-}
