--- src/main/java/com/example/openticket/global/queue/CheckQueueToken.java
@@ -1,12 +1,0 @@
-package com.example.openticket.global.queue;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface CheckQueueToken {
-    boolean consume() default true;
-}
