--- src/main/java/com/example/openticket/global/queue/QueueEntry.java
@@ -1,9 +1,0 @@
-package com.example.openticket.global.queue;
-
-public record QueueEntry(
-        Long userId,
-        String token,
-        long sequence,
-        long enteredAt
-) {
-}
