--- src/main/java/com/example/openticket/global/queue/QueuePhase.java
@@ -1,5 +1,0 @@
-package com.example.openticket.global.queue;
-
-public enum QueuePhase {
-    WAITING, ALLOWED
-}
