--- src/main/resources/lua/queue_enter.lua
@@ -1,90 +1,0 @@
--- queue_enter.lua
--- Atomic enter: idempotent check → issue token → activate or enqueue
---
--- KEYS[1] = queue:events:{eventId}:active
--- KEYS[2] = queue:events:{eventId}:waiting
--- KEYS[3] = queue:events:{eventId}:tokens
--- KEYS[4] = queue:events:{eventId}:users
--- KEYS[5] = queue:events:{eventId}:sequence
--- KEYS[6] = queue:events:registry
---
--- ARGV[1] = userId
--- ARGV[2] = maxActive
--- ARGV[3] = activeWindowSec
--- ARGV[4] = nowMs
--- ARGV[5] = newToken (UUID generated by Java)
--- ARGV[6] = eventId
---
--- Returns: {phase, token, position, remainingSeconds}
---   phase: "ALLOWED" or "WAITING"
-
-local activeKey   = KEYS[1]
-local waitingKey  = KEYS[2]
-local tokensKey   = KEYS[3]
-local usersKey    = KEYS[4]
-local sequenceKey = KEYS[5]
-local registryKey = KEYS[6]
-
-local userId          = ARGV[1]
-local maxActive       = tonumber(ARGV[2])
-local activeWindowSec = tonumber(ARGV[3])
-local nowMs           = ARGV[4]
-local newToken        = ARGV[5]
-local eventId         = ARGV[6]
-local metadataTtl     = activeWindowSec * 2
-
--- 1. Idempotent: check if user already has a token
-local existingToken = redis.call('HGET', usersKey, userId)
-if existingToken then
-    -- Check active hash
-    local activeUserId = redis.call('HGET', activeKey, existingToken)
-    if activeUserId then
-        local pttlResult = redis.call('HPTTL', activeKey, 'FIELDS', 1, existingToken)
-        local remainingMs = pttlResult[1]
-        local remainingSec = math.floor(remainingMs / 1000)
-        -- Refresh metadata TTL
-        redis.call('HEXPIRE', tokensKey, metadataTtl, 'FIELDS', 1, existingToken)
-        redis.call('HEXPIRE', usersKey, metadataTtl, 'FIELDS', 1, userId)
-        return {'ALLOWED', existingToken, '0', tostring(remainingSec)}
-    end
-
-    -- Check waiting ZSET
-    local rank = redis.call('ZRANK', waitingKey, existingToken)
-    if rank ~= false then
-        -- Refresh metadata TTL
-        redis.call('HEXPIRE', tokensKey, metadataTtl, 'FIELDS', 1, existingToken)
-        redis.call('HEXPIRE', usersKey, metadataTtl, 'FIELDS', 1, userId)
-        return {'WAITING', existingToken, tostring(rank + 1), '0'}
-    end
-
-    -- Stale: token in users hash but not in active/waiting → clean up
-    redis.call('HDEL', tokensKey, existingToken)
-    redis.call('HDEL', usersKey, userId)
-end
-
--- 2. Issue new token
-local sequence = redis.call('INCR', sequenceKey)
-local metadata = userId .. ':' .. sequence .. ':' .. nowMs
-
--- 3. Try activate
-local activeCount = redis.call('HLEN', activeKey)
-if activeCount < maxActive then
-    redis.call('HSET', activeKey, newToken, userId)
-    redis.call('HEXPIRE', activeKey, activeWindowSec, 'FIELDS', 1, newToken)
-    redis.call('HSET', tokensKey, newToken, metadata)
-    redis.call('HEXPIRE', tokensKey, metadataTtl, 'FIELDS', 1, newToken)
-    redis.call('HSET', usersKey, userId, newToken)
-    redis.call('HEXPIRE', usersKey, metadataTtl, 'FIELDS', 1, userId)
-    redis.call('SADD', registryKey, eventId)
-    return {'ALLOWED', newToken, '0', tostring(activeWindowSec)}
-end
-
--- 4. Add to waiting
-redis.call('ZADD', waitingKey, sequence, newToken)
-redis.call('HSET', tokensKey, newToken, metadata)
-redis.call('HEXPIRE', tokensKey, metadataTtl, 'FIELDS', 1, newToken)
-redis.call('HSET', usersKey, userId, newToken)
-redis.call('HEXPIRE', usersKey, metadataTtl, 'FIELDS', 1, userId)
-redis.call('SADD', registryKey, eventId)
-local rank = redis.call('ZRANK', waitingKey, newToken)
-return {'WAITING', newToken, tostring(rank + 1), '0'}
