+++ src/test/java/com/example/openticket/support/ControllerTestSupport.java
@@ -12,7 +12,7 @@ import com.example.openticket.api.service.queue.QueueService;
 import com.example.openticket.api.service.seat.SeatService;
 import com.example.openticket.domain.user.UserRepository;
 import com.example.openticket.global.auth.JwtProvider;
-import com.example.openticket.global.queue.EventQueueManager;
+import com.example.openticket.domain.queue.EventQueueManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
